+++ libcaf_core/caf/actor_system.hpp
@@ -23,6 +23,7 @@
 #include "caf/actor_traits.hpp"
 #include "caf/detail/core_export.hpp"
 #include "caf/detail/init_fun_factory.hpp"
+#include "caf/detail/private_thread_pool.hpp"
 #include "caf/detail/spawn_fwd.hpp"
 #include "caf/detail/spawnable.hpp"
 #include "caf/fwd.hpp"
@@ -530,20 +531,9 @@ public:
   actor_clock& clock() noexcept;
 
   /// Returns the number of detached actors.
-  size_t detached_actors() {
-    return detached_.load();
-  }
+  size_t detached_actors() const noexcept;
 
   /// @cond PRIVATE
-
-  /// Increases running-detached-threads-count by one.
-  void inc_detached_threads();
-
-  /// Decreases running-detached-threads-count by one.
-  void dec_detached_threads();
-
-  /// Blocks the caller until all detached threads are done.
-  void await_detached_threads();
 
   /// Calls all thread started hooks
   /// @warning must be called by thread which is about to start
@@ -634,6 +624,10 @@ public:
     return tracing_context_;
   }
 
+  detail::private_thread* acquire_private_thread();
+
+  void release_private_thread(detail::private_thread*);
+
   /// @endcond
 
 private:
@@ -699,15 +693,6 @@ private:
   /// Allows fully dynamic spawning of actors.
   strong_actor_ptr spawn_serv_;
 
-  /// Counts the number of detached actors.
-  std::atomic<size_t> detached_;
-
-  /// Guards `detached`.
-  mutable std::mutex detached_mtx_;
-
-  /// Allows waiting on specific values for `detached`.
-  mutable std::condition_variable detached_cv_;
-
   /// The system-wide, user-provided configuration.
   actor_system_config& cfg_;
 
@@ -734,6 +719,9 @@ private:
 
   /// Caches families for optional actor metrics.
   actor_metric_families_t actor_metric_families_;
+
+  /// Manages threads for detached actors.
+  detail::private_thread_pool private_threads_;
 };
 
 } // namespace caf