+++ libcaf_core/caf/detail/private_thread.hpp
@@ -7,41 +7,33 @@
 #include <atomic>
 #include <condition_variable>
 #include <mutex>
+#include <thread>
 
-#include "caf/detail/core_export.hpp"
+#include "caf/detail/private_thread_pool.hpp"
 #include "caf/fwd.hpp"
 
 namespace caf::detail {
 
-class private_thread {
+class private_thread : public private_thread_pool::node {
 public:
-  enum worker_state { active, shutdown_requested, await_resume_or_shutdown };
+  void resume(resumable* ptr);
 
-  explicit private_thread(scheduled_actor* self);
+  bool stop() override;
 
-  void run();
-
-  bool await_resume();
-
-  void resume();
-
-  void shutdown();
-
-  static void exec(private_thread* this_ptr);
-
-  void notify_self_destroyed();
-
-  void await_self_destroyed();
-
-  void start();
+  static private_thread* launch(actor_system* sys);
 
 private:
+  void run(actor_system* sys);
+
+  static void exec(actor_system* sys, private_thread* this_ptr);
+
+  std::pair<resumable*, bool> await();
+
+  std::thread thread_;
   std::mutex mtx_;
   std::condition_variable cv_;
-  std::atomic<bool> self_destroyed_;
-  std::atomic<scheduled_actor*> self_;
-  std::atomic<worker_state> state_;
-  actor_system& system_;
+  resumable* job_ = nullptr;
+  bool shutdown_ = false;
 };
 
 } // namespace caf::detail