+++ libcaf_core/src/actor_system.cpp
@@ -272,10 +272,10 @@ actor_system::actor_system(actor_system_config& cfg)
     groups_(*this),
     dummy_execution_unit_(this),
     await_actors_before_shutdown_(true),
-    detached_(0),
     cfg_(cfg),
     logger_dtor_done_(false),
-    tracing_context_(cfg.tracing_context) {
+    tracing_context_(cfg.tracing_context),
+    private_threads_(this) {
   CAF_SET_LOGGER_SYS(this);
   for (auto& hook : cfg.thread_hooks_)
     hook->init(*this);
@@ -358,6 +358,7 @@ actor_system::actor_system(actor_system_config& cfg)
   config_serv(actor_cast<strong_actor_ptr>(spawn<Flags>(config_serv_impl)));
   // Start all modules.
   registry_.start();
+  private_threads_.start();
   registry_.put("SpawnServ", spawn_serv());
   registry_.put("ConfigServ", config_serv());
   for (auto& mod : modules_)
@@ -392,7 +393,7 @@ actor_system::~actor_system() {
         ptr->stop();
       }
     }
-    await_detached_threads();
+    private_threads_.stop();
     registry_.stop();
   }
   // reset logger and wait until dtor was called
@@ -496,20 +497,8 @@ actor_clock& actor_system::clock() noexcept {
   return scheduler().clock();
 }
 
-void actor_system::inc_detached_threads() {
-  ++detached_;
-}
-
-void actor_system::dec_detached_threads() {
-  std::unique_lock<std::mutex> guard{detached_mtx_};
-  if (--detached_ == 0)
-    detached_cv_.notify_all();
-}
-
-void actor_system::await_detached_threads() {
-  std::unique_lock<std::mutex> guard{detached_mtx_};
-  while (detached_ != 0)
-    detached_cv_.wait(guard);
+size_t actor_system::detached_actors() const noexcept {
+  return private_threads_.running();
 }
 
 void actor_system::thread_started() {
@@ -541,6 +530,14 @@ actor_system::dyn_spawn_impl(const std::string& name, message& args,
   if (check_interface && !assignable(res.second, *expected_ifs))
     return sec::unexpected_actor_messaging_interface;
   return std::move(res.first);
+}
+
+detail::private_thread* actor_system::acquire_private_thread() {
+  return private_threads_.acquire();
+}
+
+void actor_system::release_private_thread(detail::private_thread* ptr) {
+  private_threads_.release(ptr);
 }
 
 } // namespace caf